+++ packages/core/src/prover.ts
@@ -1,4 +1,4 @@
-import { isWebUri } from 'valid-url'
+// import { isWebUri } from 'valid-url'
 import {
   AnonAadhaarArgs,
   AnonAadhaarProof,
@@ -53,28 +53,26 @@ export const loadZkeyChunks = async (
 }
 
 async function fetchKey(keyURL: string, maxRetries = 3): Promise<ZKArtifact> {
-  if (isWebUri(keyURL)) {
-    let attempts = 0
-    while (attempts < maxRetries) {
-      try {
-        const response = await fetch(keyURL)
-        if (!response.ok) {
-          throw new Error(
-            `Error while fetching ${retrieveFileExtension(
-              keyURL
-            )} artifacts from prover: ${response.statusText}`
-          )
-        }
-
-        const data = await response.arrayBuffer()
-        return data as Buffer
-      } catch (error) {
-        attempts++
-        if (attempts >= maxRetries) {
-          throw error
-        }
-        await new Promise(resolve => setTimeout(resolve, 1000 * attempts))
+  let attempts = 0
+  while (attempts < maxRetries) {
+    try {
+      const response = await fetch(keyURL)
+      if (!response.ok) {
+        throw new Error(
+          `Error while fetching ${retrieveFileExtension(
+            keyURL
+          )} artifacts from prover: ${response.statusText}`
+        )
       }
+
+      const data = await response.arrayBuffer()
+      return data as Buffer
+    } catch (error) {
+      attempts++
+      if (attempts >= maxRetries) {
+        throw error
+      }
+      await new Promise(resolve => setTimeout(resolve, 1000 * attempts))
     }
   }
   return keyURL