+++ CHANGELOG.md
@@ -2,8 +2,9 @@
 
 - Fix QR code verification state bug
 - Manages multiple AnonAadhaarProofs
-- Adds LaunchProveModal component that abstracts login flow
+- Adds LaunchProveModal component that abstracts login flow with button custom styles
 - Adds prover hook, to check prover state and latest proof
+- Custom artifacts links
 
 ## 1.0.1